pyproject.toml
[project]
name = "pythonformer-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==24.1.0",
    "litellm==1.81.1",
    "pydantic==2.12.5",
    "pyright==1.1.408",
    "pytest==9.0.2",
    "pyyaml==6.0.3",
    "pytest-asyncio==1.3.0",
    "python-dotenv==1.0.1",
    "rich==14.2.0",
    "nbqa>=1.7.0",
    "isort>=5.12.0",
    "ruff==0.14.13",
    "scipy==1.17.0",
    "smolagents[toolkit]==1.24.0",
]

[project.optional-dependencies]
train = [
    "axolotl",
    "datasets>=3.0.0",
    "trl>=0.12.0",
    "accelerate",
    "bitsandbytes",
    "torchvision",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
